CB-1654 Include environment ID fully in redbeams API
All redbeams API calls pertaining to databases and database servers now
require an environment ID, which may be either a CRN or name. Calls
where the environment ID is not passed as a JSON field get it from an
API query parameter, such that these resources are now scoped by
environment in the API.

Other changes:

* DatabaseConfigRepository.findByEnvironmentId now returns a set.
* DatabaseConfigService.list is renamed to findAll.
* DatabaseConfigService.get is implemented; this refactors delete logic
and forms the basis for the upcoming get API endpoint.
* DatabaseConfigService.delete for multiple databases now returns the
deleted entities, not the originally found ones.
* DatabaseV4BaseTest is introduced.
* Listing of database servers now include all of them, not just
USER_MANAGED ones.